Agrege alertas de sweetalert2

+++ src/paginas/misReservas/ContenedorReservas/contenedorReservas.jsx
@@ -2,6 +2,7 @@ import React, { useContext } from 'react'
 import { Button, Col, Stack } from 'react-bootstrap'
 import { ReservasContexto } from '../../../contexto/contexto'
 import axios from 'axios'
+import Swal from 'sweetalert2/dist/sweetalert2.js'
 
 const contenedorReservas = ({ onShowModal, Reserva }) => {
 
@@ -15,14 +16,32 @@ const contenedorReservas = ({ onShowModal, Reserva }) => {
     }
 
     const EliminarDatos = () =>{
-      axios.delete(`${UrlReservas}/${Reserva.id}`)
-      .then(response =>{
-        console.log("Reserva Eliminada con exito");
+      Swal.fire({
+        title: 'Estas seguro de que deseas eliminar esta reserva??',
+        text: "Una vez eliminada la reserva no puede ser recuperada",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Si, estoy seguro!',
+        cancelButtonText: 'No'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          axios.delete(`${UrlReservas}/${Reserva.id}`)
+          .then(response =>{
+            console.log("Reserva Eliminada con exito");
+          })
+          .catch(error => {
+            console.log("Error al eliminar la reserva");
+          })
+          TraerReservas()
+          Swal.fire(
+            'Reserva eliminada con exito!',
+            'Su reserva fue eliminada exitosamente.',
+            'success'
+          )
+        }
       })
-      .catch(error => {
-        console.log("Error al eliminar la reserva");
-      })
-      TraerReservas()
     }
 
 //Todo abajo es el contenido de las reservas, para el diseño
